+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add `payment_amount_min` parameter to `chain_metadata` entity in `/api/v1/instance` response.
+- Add information about pending invoices to instance report.
 
 ### Changed